Move DialogTriggerEditor under Editors/dialog/
