+++ src/main/cliapp/src/containers/variantsPage/VariantsTable.jsx
@@ -4,7 +4,7 @@ import { GenericDataTable } from '../../components/GenericDataTable/GenericDataT
 import { VariantService } from '../../service/VariantService';
 import { RelatedNotesEditDialog } from '../../components/RelatedNotesEditDialog';
 import { RelatedNotesReadOnlyDialog } from '../../components/RelatedNotesReadOnlyDialog';
-import { DialogTriggerEditor } from '../../components/Editors/DialogTriggerEditor';
+import { DialogTriggerEditor } from '../../components/Editors/dialog/DialogTriggerEditor';
 import { TaxonTableEditor } from '../../components/Editors/autocomplete/taxon/TaxonTableEditor';
 import { VariantTypeTableEditor } from '../../components/Editors/autocomplete/variantType/VariantTypeTableEditor';
 import { SourceGeneralConsequenceTableEditor } from '../../components/Editors/autocomplete/sourceGeneralConsequence/SourceGeneralConsequenceTableEditor';
